clarify the behavior of context stacks / parent contexts
The behavior of "context stacks" is not explained very clearly in the
documentation outside the spec, but the following text from
mustache(5) (http://mustache.github.io/mustache.5.html) is pretty
clear:

> A {{name}} tag in a basic template will try to find the name key in
> the current context. If there is no name key, the parent contexts
> will be checked recursively. If the top context is reached and the
> name key is still not found, nothing will be rendered.

In particular, when the documentation of section mentions that

> When the value is a non-empty list [..]. The context of the block
> will be set to the current item for each iteration.

or

> When the value is non-false but not a list, it will be used as the
> context for a single rendering of the block.

it should be understood that the new context becomes the "current
context", and that the previous context becomes a "parent context",
and its binding are still available.

The new tests check that these properties hold, in particular for
sections on non-object non-false values, where this was not previously
tested in the spec.