Pretty print JSON files, and terminate with newline
Having the data all on one line makes it hard for implementations that
rely on the JSON format to locate the source of a failure.

Text files that don't end with a newline can be troublesome or
unacceptable in some contexts.

The only down side is a bit of merging hassle for existing pull
requests. But that happens anyways, and of course the YAML file is
authoritative so JSON conflicts can be ignored in practice.