Fix Rakefile to add __tag__: code for ~lambda test data
This was broken when Ruby switched YAML engines from Syck to Psych
around Ruby 1.9.3. Syck was removed completely around version 2.0.
The code didn't fail, because Psych also implements the add_builtin_type
routine, but it does not recognize `!code` as a builtin type (because it
isn't; builtins look like !!str or !!map).

Instead, Psych requires a custom class to handle decoding the tagged
YAML item to Ruby.

And then, to get JSON to actually display the object, it is easiest to
just inherit from `Hash`.